requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
dj-rest-auth==2.1.11
Django==3.2.8
django-appconf==1.0.5
django-cors-headers==3.10.0
django-filter==21.1
django-imagekit==4.0.2
django-import-export==2.6.1
django-jazzmin==2.4.8
django-safedelete==1.1.0
django-taggit==1.5.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-generators==0.5.0
et-xmlfile==1.1.0
idna==3.3
itypes==1.2.0
Jinja2
MarkupPy==1.14
MarkupSafe
odfpy==1.4.1
openpyxl==3.0.9
pilkit==2.0
Pillow==8.4.0
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
SpeechRecognition
# pyaudio
moviepy
bs4==0.0.1
html2text==2020.1.16
icecream==2.1.1
pyttsx3==2.90
ffmpeg-python==0.2.0
